cmd: prevent exiting before stdout flush
Currently the command outputs to stdout by calling
`console.log()`, and if the parsed tap input contains
failures it will forcibly exit the process. This can cause
the process to exit before the stdout stream is flushed.

Instead of `process.exit(1)`, this sets the `exitCode`
and allows the process to exit gracefully.

Potentially related to #50 and #59